Move TextTransform classes into text-transform directory
