+++ app/writer/text-transform/ReactTextTransform.spec.js
