+++ app/writer/text-transform/TextTransform.spec.js
